New nav styles WIP
