+++ scripts/pages/about.page.js
@@ -1,5 +1,6 @@
 import 'styles/about.scss';
 
+import Nav from 'components/nav.component.js';
 import smoothScroll from 'utils/smoothScroll';
 import { calculateOffsets, scrollDocument } from 'utils/fixedScroll';
 import _ from 'lodash';
@@ -58,3 +59,5 @@ const _setEventListeners = () => {
 
 _calculateOffsets();
 _setEventListeners();
+const nav = new Nav();
+nav.onCreated();
