+++ styles/components/shared/nav.scss
@@ -10,7 +10,7 @@
   nav {
     height: $nav-height;
 
-    > li {
+    > .nav-item {
       position: relative;
       display: inline-flex;
       align-items: center;
@@ -21,9 +21,16 @@
       letter-spacing: 0.5px;
       color: $charcoal-grey;
 
+      > .nav-link {
+        position: relative;
+        display: flex;
+        align-items: center;
+        height: 100%;
+      }
+
       &.-has-menu {
 
-        &:hover .submenu {
+        &:hover .nav-submenu {
           display: block;
         }
       }
@@ -35,15 +42,75 @@
       &:last-child {
         padding-right: 0;
       }
+
+      &.-selected > .nav-link:before {
+        content: "";
+        position: absolute;
+        left: 0;
+        bottom: 0;
+        width: 100%;
+        height: 6px;
+        background-color: $medium-pink;
+      }
     }
 
-    .submenu {
+    .nav-submenu {
       display: none;
-      position: absolute;
-      top: $nav-height;
-      left: 14px;
-      min-width: 200px;
+      position: fixed;
+      top: $nav-height - 11;
+      background-color: $white;
+      box-shadow: 0 10px 20px 0 rgba($charcoal-grey, .2), 0 0 20px 0 rgba($charcoal-grey, .18);
+      border: solid 1px rgba($charcoal-grey, .2);
+      z-index: 1;
+
+      > .nav-item {
+        display: flex;
+        align-items: center;
+        font-weight: normal;
+        padding: 0 16px;
+        height: 48px;
+
+        &:not(:first-child) {
+          border-top: solid 1px rgba($charcoal-grey, .2);
+        }
+
+        &:hover {
+          background-color: $egg-shell;
+        }
+      }
+
+      &:after,
+      &:before {
+        content: "";
+        position: absolute;
+        bottom: 100%;
+        left: 11px;
+        border: solid transparent;
+        height: 0;
+        width: 0;
+      }
+
+      &:after {
+        border-bottom-color: $white;
+        border-width: 5px;
+        margin-left: -5px;
+      }
+
+      &:before {
+        border-bottom-color: rgba($charcoal-grey, .2);
+        border-width: 6px;
+        margin-left: -6px;
+      }
     }
+
+    &.-right {
+      text-align: right;
+    }
+  }
+
+  &.-have-background {
+    background-color: $white;
+    box-shadow: 0 10px 20px 0 rgba($charcoal-grey, 0.2);
   }
 }
